+++ sway-lsp/src/capabilities/hover.rs
@@ -4,7 +4,7 @@ use crate::{
         token::{Token, TypedAstToken},
     },
     utils::{
-        attributes::doc_attributes, common::get_range_from_span, markdown, markup::Markup,
+        attributes::doc_attributes, markdown, markup::Markup, token::get_range_from_span,
         token::to_ident_key,
     },
 };
