+++ sway-lsp/src/capabilities/inlay_hints.rs
@@ -1,6 +1,7 @@
 use crate::{
-    core::{config::InlayHintsConfig, session::Session, token::TypedAstToken},
-    utils::common::get_range_from_span,
+    config::InlayHintsConfig,
+    core::{session::Session, token::TypedAstToken},
+    utils::token::get_range_from_span,
 };
 use std::sync::Arc;
 use sway_core::{language::ty::TyDeclaration, type_system::TypeInfo};
@@ -66,7 +67,7 @@ pub(crate) fn inlay_hints(
             }
         })
         .map(|var| {
-            let range = crate::utils::common::get_range_from_span(&var.name.span());
+            let range = crate::utils::token::get_range_from_span(&var.name.span());
             let kind = InlayKind::TypeHint;
             let label = format!("{}", type_engine.help_out(var.type_ascription));
             let inlay_hint = InlayHint { range, kind, label };
