+++ sway-lsp/src/config.rs
