+++ sway-lsp/src/core/session.rs
@@ -4,7 +4,7 @@ use crate::{
         formatting::get_page_text_edit,
         runnable::{Runnable, RunnableType},
     },
-    core::{dependency, document::TextDocument, token_map::TokenMap},
+    core::{document::TextDocument, token_map::TokenMap},
     error::{DocumentError, LanguageServerError},
     traverse,
     utils::{self, sync::SyncWorkspace},
@@ -171,11 +171,14 @@ impl Session {
                     capabilities::diagnostic::get_diagnostics(&ast_res.warnings, &ast_res.errors);
             } else {
                 // Collect tokens from dependencies and the standard library prelude.
-                self.parse_ast_to_tokens(parse_program, dependency::collect_parsed_declaration);
+                self.parse_ast_to_tokens(
+                    parse_program,
+                    traverse::dependency::collect_parsed_declaration,
+                );
 
                 self.parse_ast_to_typed_tokens(
                     typed_program,
-                    dependency::collect_typed_declaration,
+                    traverse::dependency::collect_typed_declaration,
                 );
             }
         }
@@ -249,7 +252,7 @@ impl Session {
         let token_ranges = self
             .token_map
             .all_references_of_token(&token, &self.type_engine.read())
-            .map(|(ident, _)| utils::common::get_range_from_span(&ident.span()))
+            .map(|(ident, _)| utils::token::get_range_from_span(&ident.span()))
             .collect();
 
         Some(token_ranges)
@@ -264,7 +267,7 @@ impl Session {
             .token_at_position(&uri, position)
             .and_then(|(_, token)| token.declared_token_ident(&self.type_engine.read()))
             .and_then(|decl_ident| {
-                let range = utils::common::get_range_from_span(&decl_ident.span());
+                let range = utils::token::get_range_from_span(&decl_ident.span());
                 decl_ident.span().path().and_then(|path| {
                     // We use ok() here because we don't care about propagating the error from from_file_path
                     Url::from_file_path(path.as_ref()).ok().and_then(|url| {
@@ -361,7 +364,7 @@ impl Session {
             ref main_function, ..
         } = typed_program.kind
         {
-            let main_fn_location = utils::common::get_range_from_span(&main_function.name.span());
+            let main_fn_location = utils::token::get_range_from_span(&main_function.name.span());
             let runnable = Runnable::new(main_fn_location, typed_program.kind.tree_type());
             self.runnables.insert(RunnableType::MainFn, runnable);
         }
