+++ sway-lsp/src/traverse/dependency.rs
