+++ sway-lsp/src/traverse/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod dependency;
 pub(crate) mod parse_tree;
 pub(crate) mod typed_tree;
