+++ sway-lsp/src/utils/mod.rs
@@ -1,5 +1,4 @@
 pub(crate) mod attributes;
-pub(crate) mod common;
 pub mod debug;
 pub mod markdown;
 pub(crate) mod markup;
